Switch to a container-based development environment
To produce a container image using Nix, I had to re-package the Python
project in a way that satisfied Nix. I found myself falling down a
rabbit hole, re-packaging the Python project using Poetry and trying to
use poetry2nix to turn the project into a Nix derivation. We use a lot
of Python dependencies that depend on C libraries (like `pyparsing`,
`python-magic`, `pillow`, etc.), and I had to patch each one to fit the
Nix way of making shared libraries available to packages.

This research was OK because we intentionally baked in some time for
exploring Nix, but I began to feel like my wheels were spinning, and I
wouldn’t arrive at a workable solution in time for our early May
deadline.

I decided to pivot to a more traditional approach using Docker and
Docker Compose to manage the development environment and produce the
container image.

This approach feels a lot more straightforward. As a result, I can
deliver faster, and we better minimize the divergence between
development and production. However, we lose the reproducible builds and
native performance of Nix. Also, we now have to deal with some remote
debugger trickery with PyCharm.

I made the following changes to support this switch:

- Replace `flake.nix` with a `Dockerfile` for building a container image
  containing all of the CLI’s dependencies (e.g., the AWS CLI, dcm2niix,
  and the Flywheel CLI).
- Package the Python project using `setuptools` to wrap the application
  in an `image-deid-etl` executable.
- Add service definitions for `docker-compose.yml` that mimic how we’d
  run the container on AWS.
- Wrap the `upload2fw` step in a fake user home directory so that the
  Flywheel CLI can consume credentials from the environment.
- Install the `psycopg2` and PostgreSQL client libraries and update the
  database schema to support PostgreSQL.