Revert default_tags and tag all resources by-hand
There is a known issue with the Terraform provider where using
default_tags will always trigger a diff if you specify the same tags at
the resource level:

https://github.com/hashicorp/terraform-provider-aws/issues/18311

I tried to use default_tags in this project to make the Terraform code
more readable. But, we're running into a problem with Azavea's
PostgreSQL module because it duplicates the Project and Environment tags
at the resource level, causing a noisy diff.